CP-53554: Split plug xenopsd atomic into attach/activate
This consists of two parts, splitting the attach_and_activate into
functionally equivalent attach and activate functions, and splitting the
VBD_plug atomic itself's behaviour into two new atomics, VBD_attach and
VBD_activate.

If the new xenopsd_vbd_plug_unplug_legacy flag is true, the only
difference will be that VBD_plug calls attach and activate sequentially
instead of attach_and_activate.

If xenopsd_vbd_plug_unplug_legacy is false, the VBD_attach and
VBD_activate atomics will be used in place of VBD_plug in all places
that it is used. This should still be functionally equivalent.

The purpose of this change is to allow optimisations in this area as
there are some situations where they do not need to be called at the
same time. For example VBD_attach could be called outside of VM migrate
downtime to reduce the overall downtime.

Signed-off-by: Steven Woods <steven.woods@cloud.com>