Remove network attached to protected vms check
This check blocks PIF unplugs on networks with protected VMs and
similarly blocks host enables with HA if any PIFs are unplugged on these
networks. This check is unnecessary as PIFs are plugged automatically
when needed. This also causes HA to erroneously block exiting maintenance mode
if a host has been rebooted when there are VLAN PIFs (as they become
unattached after reboot).